temporarily simplify installation
